requirements.txt
python-telegram-bot==21.4
python-telegram-bot[job-queue]==21.4
Flask==3.0.3
psycopg[binary]==3.2.2
psycopg_pool==3.2.1
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.27.0
asgiref==3.8.1
# New additions for upgrades
redis==5.0.1  # For caching layer
aioredis==2.0.1  # Async Redis support
APScheduler==3.10.4  # For scheduled tasks
pydantic==2.5.0  # Data validation
